requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
streamlit>=1.30.0
groq>=0.9.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
networkx>=3.1
sqlalchemy>=2.0.0
PyMuPDF>=1.23.0
numpy>=1.24.0
requests>=2.31.0
plotly>=5.18.0